Add user authentication routes and controllers
